requirements.txt
adafruit-ampy==1.1.0
altgraph==0.17.4
argcomplete==3.5.1
bitarray==2.9.3
bitstring==4.2.3
cffi==1.17.1
click==8.1.7
cryptography==43.0.3
ecdsa==0.19.0
esptool==4.8.1
importlib_metadata==8.5.0
intelhex==2.3.0
jedi==0.19.2
macholib==1.16.3
packaging==24.2
parso==0.8.4
pycparser==2.22
Pygments==2.18.0
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2024.10
PyQt5==5.15.11
PyQt5-Qt5==5.15.15
PyQt5_sip==12.15.0
PySide6==6.4.2
PySide6-plugins==6.4.2.2.3
PySide6-Qt6==6.4.3
PySide6-tools==6.4.2.3.3
PySide6_sip==13.8.0
pyserial==3.5
python-dotenv==1.0.1
PyYAML==6.0.2
QScintilla==2.14.1
qt6-applications==6.4.3.2.3
qt6-tools==6.4.3.1.3
reedsolo==1.7.0
six==1.16.0
zipp==3.21.0
